requirements.txt
asgiref==3.4.1
certifi==2021.5.30
Django==3.2.8
django-filebrowser==3.14.1
django-grappelli==2.15.1
django-tinymce4-lite==1.8.0
jsmin==3.0.0
Pillow==8.3.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
wincertstore==0.2
